renamed-component-additional-info
