+++ src/components/UI/Select.tsx
@@ -21,10 +21,9 @@ export default function BasicSelect(props: BasicSelectProps) {
   const handleOpen = () => {
     setOpen(!open);
   };
-  // console.log(open);
   return (
     <Box sx={{ minWidth: 120 }}>
-      <label className="block mb-2">{props.label}</label>
+      <label className="block mb-2 body-3 text-tertiary">{props.label}</label>
       <FormControl
         fullWidth
         sx={{
